tech(api): move get-next-challenge-for-demo in evaluation bounded context
