+++ api/src/evaluation/domain/usecases/get-next-challenge-for-demo.js
@@ -1,0 +1,26 @@
+import { AssessmentEndedError } from '../../../shared/domain/errors.js';
+import { _ } from '../../../shared/infrastructure/utils/lodash-utils.js';
+
+const getNextChallengeForDemo = async function ({
+  assessment,
+  answerRepository,
+  challengeRepository,
+  courseRepository,
+}) {
+  const course = await courseRepository.get(assessment.courseId);
+  const answers = await answerRepository.findByAssessment(assessment.id);
+  const nextChallengeId = _selectNextChallengeId(course, answers);
+  if (!nextChallengeId) {
+    throw new AssessmentEndedError();
+  }
+  return challengeRepository.get(nextChallengeId);
+};
+
+export { getNextChallengeForDemo };
+
+function _selectNextChallengeId(course, answers) {
+  const courseChallengeIds = course.challenges;
+  const answeredChallengeIds = _.map(answers, 'challengeId');
+
+  return _(courseChallengeIds).difference(answeredChallengeIds).first();
+}
